MessageContainerを新規作成し、Pages/indexからMessagesの要素たちを分離
