+++ components/MessageList.vue
@@ -1,7 +1,5 @@
 <template lang="pug">
-  .scroll-y(ref="messageListCard")
-    v-btn.scroll-btn(absolute fab small color="primary" @click='scrollToEnd')
-      v-icon(dark) keyboard_arrow_down
+  div
     v-list(two-line)
       div(
         v-for="(message, index) in messages"
@@ -23,22 +21,9 @@ export default {
   },
   computed: {
     ...mapState('messages', ['messages']),
-    ...mapState('rooms', ['selectedRoomID']),
-  },
-  watch: {
-    selectedRoomID: function() {
-      this.$store.dispatch('messages/stopListener')
-      this.$store.dispatch('messages/startListener')
-    },
-  },
-  mounted() {
-    this.$store.dispatch('messages/startListener')
   },
   updated() {
-    this.scrollToEnd()
-  },
-  destroyed() {
-    this.$store.dispatch('messages/stopListener')
+    this.$emit('new-message')
   },
   methods: {
     UNIXtimeToDate(UNIXtime) {
@@ -48,17 +33,6 @@ export default {
       const getDate = message => this.UNIXtimeToDate(message.data.timestamp)
       return getDate(currentMessage) !== getDate(prevMessage)
     },
-    scrollToEnd() {
-      const container = this.$refs.messageListCard
-      container.scrollTop = container.scrollHeight
-    },
   },
 }
 </script>
-
-<style scoped lang="sass">
-  .scroll-btn
-    top: 0pt
-    right: 0pt
-    margin: 20pt
-</style>
